Refactor commands to files
